+++ flashinfer/fused_moe/cute_dsl/blackwell_sm12x/moe_dynamic_kernel.py
@@ -4763,4 +4763,66 @@ class MoEDynamicKernelBackend:
         return
 
 
-__all__ = ["MoEDynamicKernelBackend"]
+class MoEDynamicKernelSilu(MoEDynamicKernelBackend):
+    def __init__(
+        self,
+        sf_vec_size: int,
+        mma_tiler_mn: Tuple[int, int],
+        *,
+        fast_math: bool = False,
+        dynamic_down_scale: bool = False,
+        share_input_across_experts: bool = False,
+        deterministic_output: bool = False,
+        swap_ab: bool = False,
+        quant_recipe: str = "nvfp4",
+        swiglu_limit: float | None = None,
+        swiglu_alpha: float | None = None,
+        swiglu_beta: float | None = None,
+    ):
+        super().__init__(
+            sf_vec_size,
+            mma_tiler_mn,
+            fast_math=fast_math,
+            activation="silu",
+            dynamic_down_scale=dynamic_down_scale,
+            share_input_across_experts=share_input_across_experts,
+            deterministic_output=deterministic_output,
+            swap_ab=swap_ab,
+            quant_recipe=quant_recipe,
+            swiglu_limit=swiglu_limit,
+            swiglu_alpha=swiglu_alpha,
+            swiglu_beta=swiglu_beta,
+        )
+
+
+class MoEDynamicKernelRelu2(MoEDynamicKernelBackend):
+    def __init__(
+        self,
+        sf_vec_size: int,
+        mma_tiler_mn: Tuple[int, int],
+        *,
+        fast_math: bool = False,
+        dynamic_down_scale: bool = False,
+        share_input_across_experts: bool = False,
+        deterministic_output: bool = False,
+        swap_ab: bool = False,
+        quant_recipe: str = "nvfp4",
+    ):
+        super().__init__(
+            sf_vec_size,
+            mma_tiler_mn,
+            fast_math=fast_math,
+            activation="relu2",
+            dynamic_down_scale=dynamic_down_scale,
+            share_input_across_experts=share_input_across_experts,
+            deterministic_output=deterministic_output,
+            swap_ab=swap_ab,
+            quant_recipe=quant_recipe,
+        )
+
+
+__all__ = [
+    "MoEDynamicKernelBackend",
+    "MoEDynamicKernelSilu",
+    "MoEDynamicKernelRelu2",
+]
