+++ tests/moe/test_b12x_fused_moe.py
@@ -1214,151 +1214,6 @@ class TestB12xFunctional:
             f"(atol={atol:.4f}, tokens={num_tokens})"
         )
 
-    @pytest.mark.parametrize("layered", [False, True])
-    def test_w4a16_activation_amax_calibration(self, layered: bool):
-        """W4A16 ``collect_activation_amax`` hook (backport of b12x d53a1b2).
-
-        Asserts the opt-in calibration hook (a) does not perturb the forward
-        output (bit-for-bit identical to the uncalibrated run) and (b) collects
-        the per-expert max-abs of the FC1-input activations (slot 0) and the
-        FC2-input activations (slot 1), matching a torch reference oracle.
-        """
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x.moe_dispatch import (
-            _allocate_sm120_w4a16_workspace,
-            _get_w4a16_packed_weights,
-        )
-        from flashinfer.fused_moe.cute_dsl.blackwell_sm12x.moe_w4a16_kernel import (
-            run_w4a16_moe,
-        )
-
-        device = torch.device("cuda")
-        hidden_size, intermediate_size = 256, 512
-        num_experts, top_k = 8, 2
-        num_tokens = 64
-        activation = "silu"
-        tensors = create_moe_tensors(
-            num_tokens=num_tokens,
-            hidden_size=hidden_size,
-            intermediate_size=intermediate_size,
-            num_experts=num_experts,
-            num_local_experts=num_experts,
-            top_k=top_k,
-            seed=20260616,
-        )
-        a_input = tensors["x_bf16"]
-        topk_ids = tensors["token_selected_experts"]
-        topk_weights = tensors["token_final_scales"]
-
-        prepared = _get_w4a16_packed_weights(
-            w1_weight=tensors["w1_weight"],
-            w1_weight_sf=tensors["w1_weight_sf"],
-            w1_alpha=tensors["w1_alpha"],
-            w2_weight=tensors["w2_weight"],
-            w2_weight_sf=tensors["w2_weight_sf"],
-            w2_alpha=tensors["w2_alpha"],
-            activation=activation,
-            params_dtype=a_input.dtype,
-            source_format="modelopt",
-        )
-
-        def fresh_workspace():
-            return _allocate_sm120_w4a16_workspace(
-                state_E=num_experts,
-                weight_E=num_experts,
-                routed_rows=num_tokens * top_k,
-                k=hidden_size,
-                n=intermediate_size,
-                num_topk=top_k,
-                device=device,
-                activation=activation,
-            )
-
-        def run(ws, output, activation_amax=None, layer_idx=None):
-            return run_w4a16_moe(
-                a_input,
-                prepared,
-                topk_weights,
-                topk_ids,
-                activation=activation,
-                intermediate_cache13=ws.intermediate_cache13,
-                intermediate_cache2=ws.intermediate_cache2,
-                output=output,
-                fc1_c_tmp=ws.fc1_c_tmp,
-                fc2_c_tmp=ws.fc2_c_tmp,
-                packed_route_indices=ws.packed_route_indices,
-                block_expert_ids=ws.block_expert_ids,
-                packed_route_count=ws.packed_route_count,
-                expert_offsets=ws.expert_offsets,
-                expert_map=ws.expert_map,
-                activation_amax=activation_amax,
-                layer_idx=layer_idx,
-            )
-
-        # Baseline (calibration OFF).
-        ws_base = fresh_workspace()
-        out_base = torch.empty(
-            (num_tokens, hidden_size), dtype=a_input.dtype, device=device
-        )
-        run(ws_base, out_base)
-        torch.cuda.synchronize()
-        out_base = out_base.clone()
-
-        # Calibration ON. Use a 3D [num_layers, num_experts, 2] tensor in the
-        # layered variant (selecting layer 1) and a 2D [num_experts, 2] tensor
-        # otherwise, exercising both validated shapes. A sentinel layer is filled
-        # with a large value to confirm only the selected layer is touched.
-        ws_cal = fresh_workspace()
-        out_cal = torch.empty(
-            (num_tokens, hidden_size), dtype=a_input.dtype, device=device
-        )
-        if layered:
-            activation_amax = torch.zeros(
-                (2, num_experts, 2), dtype=torch.float32, device=device
-            )
-            activation_amax[0].fill_(123.0)
-            layer_idx = 1
-        else:
-            activation_amax = torch.zeros(
-                (num_experts, 2), dtype=torch.float32, device=device
-            )
-            layer_idx = None
-        run(ws_cal, out_cal, activation_amax=activation_amax, layer_idx=layer_idx)
-        torch.cuda.synchronize()
-        activated = ws_cal.intermediate_cache2.view(-1)[
-            : num_tokens * top_k * intermediate_size
-        ].view(num_tokens * top_k, intermediate_size)
-        activated = activated.clone()
-
-        # (a) The hook must not perturb the forward output.
-        torch.testing.assert_close(out_cal, out_base, atol=0.0, rtol=0.0)
-
-        # (b) The collected amax must match a torch reference of the per-expert
-        #     FC1-input (slot 0) and FC2-input (slot 1) max-abs activations.
-        collected = activation_amax[1] if layered else activation_amax
-        expected_amax = torch.zeros(
-            (num_experts, 2), dtype=torch.float32, device=device
-        )
-        flat_ids = topk_ids.reshape(-1)
-        route_tokens = torch.arange(num_tokens, device=device).repeat_interleave(top_k)
-        route_indices = torch.arange(num_tokens * top_k, device=device)
-        for expert in range(num_experts):
-            mask = flat_ids == expert
-            if bool(mask.any().item()):
-                expected_amax[expert, 0] = (
-                    a_input[route_tokens[mask]].float().abs().max()
-                )
-                expected_amax[expert, 1] = (
-                    activated[route_indices[mask]].float().abs().max()
-                )
-        torch.testing.assert_close(collected, expected_amax, atol=0.0, rtol=0.0)
-        if layered:
-            torch.testing.assert_close(
-                activation_amax[0],
-                torch.full_like(activation_amax[0], 123.0),
-                atol=0.0,
-                rtol=0.0,
-            )
-
     def test_w4a16_static_accuracy_intermediate_not_128_aligned(self):
         """W4A16 must handle n=64 mod 128 without crossing gate/up tiles."""
         from flashinfer import b12x_fused_moe
